add ALL the tests\!

+++ test/test_helper.coffee
@@ -2,15 +2,22 @@ MongoClient = require('mongodb').MongoClient
 
 config = require '../config'
 
-before (done) ->
-  env = process.env.NODE_ENV
+env = process.env.NODE_ENV
+if env != 'test'
+  console.log '** NODE_ENV is not test! **'
+  console.log 'Exiting now. No test performed.'
+  process.exit(1)
+
+dropDB = (cb) ->
   MongoClient.connect config[env]['db']['url'], (err, db) ->
     if err
       throw err
     db.dropDatabase (err, result) ->
+      console.log 'dropDatabase', err, result
       if err
         throw err
-      done()
+      cb()
 
 module.exports =
   host: 'http://localhost:3000'
+  dropDB: dropDB
